Fix: 15 SDK bugs + per-game hash_eagle layout + new-game docs
SDK CORE (neogeolib.c)
- Bug 1: displayCreditP1/P2 units digit wrote d1 instead of d0
- Bug 2: displayCreditP2 used same VRAM address as P1 (0x7088 -> 0x7288)
- Bug 3: setFIXDATA palette shift 11 -> 12 (bits 15-12 per FIX cell spec)

2D ENGINE
- Bug 4: ng_char_draws_before depth sort reversed; ay < by -> ay > by
- Bug 5: particle eviction drop_pri=0xff; no uint8_t can exceed 0xff -> 0
- Bug 6: palette upload base truncated to uint16_t; PALETTES=0x400000 -> uintptr_t
- Bug 7: ng_camera_apply applied bounds after update; swap order so first
  frame uses correct bounds
- Bug 8: ng_sprite_group upload() chain strip SCB3 was 0x0040, inconsistent
  with update_transform/flush which write 0x0040|activeRows
- Bug 9: NG_IMPACT_HEAVY flash_red immediately overwrote flash_white in same
  palette slot; remove the immediate red call
- Bug 10: ng_dbg_particles uint16_t in debug path, uint8_t in release;
  unified to uint16_t in both .c and .h
- Bug 11: ng_game_engine_init/frame did not wire render_queue, particles,
  palette_fx, feedback, depthfx; added includes and init/update calls

DEMO
- Bug 12: DemoShowScreenFn typedef duplicated in demo.h and demo_screen.h;
  removed from demo_screen.h, replaced with #include "demo.h"
- Bug 13: demo_screen, demo_2d_engine, demo_3d not in GAME_SCENES; added
  to games/demo/game.mk
- Bug 14: particle eviction demo broken by Bug 5; fixed by Bug 5 fix
- Bug 15: FIX layer text never appeared; ngfix_init and ngfix_clear did not
  call setsfix() so BRDFIX stayed 0 (BIOS S ROM) after boot and clearFix

HASH EAGLE
- Each game now gets hash_eagle/<GAME>/neogeo.xml (isolated per-game XML)
- HASHPATH = hash_eagle/<GAME>:hash_eagle:hash (game dir takes priority)
- dist launchers use per-game hashpath for both .bat and .sh
- Existing neogeo.xml moved to hash_eagle/demo/neogeo.xml

DOCS
- docs/ADDING_A_GAME.md: step-by-step guide for adding a new game

+++ sdk/2d_engine/ng_render_queue.c
@@ -130,8 +130,8 @@ void NEOGEO_USER ng_render_queue_flush(void)
     /* --- Apply palette uploads --- */
     for (i = 0; i < ng_rq_pal_count; i++) {
         const NGPaletteUpload *p = &ng_rq_palettes[i];
-        uint16_t base = (uint16_t)(PALETTES + (uint32_t)p->palette_addr * 2u);
-        load_palettes((uint16_t *)p->data, (uintptr_t)base);
+        uintptr_t base = (uintptr_t)((uint32_t)PALETTES + (uint32_t)p->palette_addr * 2u);
+        load_palettes((uint16_t *)p->data, base);
     }
     ng_rq_pal_count = 0;
 }
